+++ src/exporter/xlsx.rs
@@ -488,7 +488,7 @@ mod tests {
 
     use crate::{
         ExporterStep, ImporterStep, ReadFrom, WriteAs,
-        importer::{ImportRunConfiguration, Importer, xlsx::ImportSpreadsheet},
+        importer::{GenericImportConfiguration, Importer, xlsx::ImportSpreadsheet},
         test_util::compare_graphs,
         util::example_generator,
     };
@@ -549,8 +549,7 @@ mod tests {
             module: crate::ReadFrom::Xlsx(importer),
             path: path.to_path_buf(),
             description: None,
-            extensions: None,
-            root_name: None,
+            generic_config: None,
         };
         let mut updates = orig_import_step.execute(None).unwrap();
         let mut original_graph = AnnotationGraph::with_default_graphstorages(false).unwrap();
@@ -579,8 +578,7 @@ mod tests {
             module: crate::ReadFrom::Xlsx(importer),
             path: output_dir.clone(),
             description: None,
-            extensions: None,
-            root_name: None,
+            generic_config: None,
         };
         let mut updates = second_import_step.execute(None).unwrap();
         let mut written_graph = AnnotationGraph::with_default_graphstorages(false).unwrap();
@@ -607,8 +605,7 @@ mod tests {
             module: importer,
             path: path.to_path_buf(),
             description: None,
-            extensions: None,
-            root_name: None,
+            generic_config: None,
         };
         let mut updates = orig_import_step.execute(None).unwrap();
         let mut original_graph = AnnotationGraph::with_default_graphstorages(false).unwrap();
@@ -637,8 +634,7 @@ mod tests {
             module: crate::ReadFrom::Xlsx(importer),
             path: output_dir.clone(),
             description: None,
-            extensions: None,
-            root_name: None,
+            generic_config: None,
         };
         let mut updates = second_import_step.execute(None).unwrap();
 
@@ -684,8 +680,7 @@ mod tests {
             module: importer,
             path: path.to_path_buf(),
             description: None,
-            extensions: None,
-            root_name: None,
+            generic_config: None,
         };
 
         let mut updates = first_import_step.execute(None).unwrap();
@@ -714,8 +709,7 @@ mod tests {
             module: crate::ReadFrom::Xlsx(importer),
             path: output_dir.clone(),
             description: None,
-            extensions: None,
-            root_name: None,
+            generic_config: None,
         };
         let mut updates = second_import_step.execute(None).unwrap();
 
@@ -758,8 +752,7 @@ mod tests {
             module: importer,
             path: path.to_path_buf(),
             description: None,
-            extensions: None,
-            root_name: None,
+            generic_config: None,
         };
         let mut updates = orig_import_step.execute(None).unwrap();
         let mut original_graph = AnnotationGraph::with_default_graphstorages(false).unwrap();
@@ -790,8 +783,7 @@ mod tests {
             module: crate::ReadFrom::Xlsx(importer),
             path: output_dir.clone(),
             description: None,
-            extensions: None,
-            root_name: None,
+            generic_config: None,
         };
         let e = second_import_step.execute(None);
         assert!(e.is_ok(), "Error re-importing: {:?}", e.err().unwrap());
@@ -851,8 +843,7 @@ mod tests {
             module: importer,
             path: corpus_dir.path().to_path_buf(),
             description: None,
-            extensions: None,
-            root_name: None,
+            generic_config: None,
         };
         let mut updates = orig_import_step.execute(None).unwrap();
         let mut graph = AnnotationGraph::with_default_graphstorages(false).unwrap();
@@ -903,8 +894,7 @@ mod tests {
             module: importer,
             path: corpus_dir.path().to_path_buf(),
             description: None,
-            extensions: None,
-            root_name: None,
+            generic_config: None,
         };
         let mut updates = orig_import_step.execute(None).unwrap();
         let mut graph = AnnotationGraph::with_default_graphstorages(false).unwrap();
@@ -957,8 +947,7 @@ mod tests {
             module: importer,
             path: corpus_dir.path().to_path_buf(),
             description: Some("custom-xlsx-export-id".to_string()),
-            extensions: None,
-            root_name: None,
+            generic_config: None,
         };
         let mut updates = orig_import_step.execute(None).unwrap();
         let mut graph = AnnotationGraph::with_default_graphstorages(false).unwrap();
@@ -1015,7 +1004,7 @@ mod tests {
                     module_name: "test_import".to_string(),
                     path: None,
                 },
-                ImportRunConfiguration::new_with_default_extensions(&importer),
+                GenericImportConfiguration::new_with_default_extensions(&importer),
                 None,
             )
             .unwrap();