+++ src/importer/saltxml.rs
@@ -4,7 +4,7 @@ use graphannis::update::GraphUpdate;
 use roxmltree::Node;
 use serde::{Deserialize, Serialize};
 
-use crate::{importer::ImportRunConfiguration, progress::ProgressReporter};
+use crate::{importer::GenericImportConfiguration, progress::ProgressReporter};
 
 use super::Importer;
 
@@ -38,7 +38,7 @@ impl Importer for ImportSaltXml {
         &self,
         input_path: &std::path::Path,
         step_id: crate::StepID,
-        _config: ImportRunConfiguration,
+        _config: GenericImportConfiguration,
         tx: Option<crate::workflow::StatusSender>,
     ) -> Result<graphannis::update::GraphUpdate, Box<dyn std::error::Error>> {
         let mut updates = GraphUpdate::new();