+++ src/importer/xml/mod.rs
@@ -13,7 +13,7 @@ use xml::{EventReader, ParserConfig};
 use crate::{
     StepID,
     error::{AnnattoError, Result},
-    importer::ImportRunConfiguration,
+    importer::GenericImportConfiguration,
     progress::ProgressReporter,
     util::graphupdate::import_corpus_graph_from_files,
 };
@@ -44,7 +44,7 @@ impl Importer for ImportXML {
         &self,
         input_path: &std::path::Path,
         step_id: crate::StepID,
-        config: ImportRunConfiguration,
+        config: GenericImportConfiguration,
         tx: Option<crate::workflow::StatusSender>,
     ) -> std::result::Result<GraphUpdate, Box<dyn std::error::Error>> {
         let mut update = GraphUpdate::default();